requirements.txt
anthropic==0.62.0
streamlit==1.48.1
streamlit-webrtc==0.63.4
streamlit-autorefresh==1.0.1
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
ultralytics==8.3.170
deepface==0.0.95
librosa==0.11.0
reportlab==4.4.3
matplotlib==3.10.5
plyer==2.1.0
pyyaml==6.0.2
pandas==2.3.1
numpy==1.26.4
scipy==1.16.1
scikit-learn==1.7.1
sqlalchemy==2.0.43
python-dotenv==1.1.1
tf-keras==2.19.0
SpeechRecognition==3.14.3
resemblyzer==0.1.4
